Remove nonUi
