+++ server/lib/cipher.js
@@ -1,6 +1,6 @@
 const crypto = require('crypto')
 const algorithm = 'aes256'
-const { passphrase } = require('../lib/config/nonUi').getConfig()
+const { passphrase } = require('../lib/config').getConfig()
 
 module.exports = function(text) {
   const myCipher = crypto.createCipher(algorithm, passphrase)
